+++ examples/source/Geothermal_Example.h
@@ -8,7 +8,7 @@
 namespace csmp {
 
 const size_t DIM(2U);
-template<size_t> class Model;
+template<uint32_t> class Model;
 
 class Geothermal_Example : public Example {
   public:
