+++ source/includes/discretisation/Quadrilaterator.h
@@ -6,7 +6,7 @@
 
 namespace csmp {
 
-template<size_t> class VSet;
+template<uint32_t> class VSet;
 
 /**
