+++ source/includes/integration/finite_elements/IsoparametricLinearQuadrilateral.h
@@ -9,43 +9,43 @@ class IsoparametricLinearQuadrilateral : public FiniteElement {
 
 public:
 
-    explicit IsoparametricLinearQuadrilateral( size_t dimensions = 2 );
+    explicit IsoparametricLinearQuadrilateral( uint32_t dimensions = 2 );
     ~IsoparametricLinearQuadrilateral();
 
     virtual double    Volume();
     virtual double    AspectRatio();
     virtual double    InnerRadius();
     virtual void        EdgeLengths( std::vector<double>& vec );
-    virtual void        NodesOfSegment( size_t segm_id, std::vector<size_t>& snids ) const;
-    virtual void        NodesOfFace( size_t face_id, std::vector<size_t>& fnids ) const;
-    virtual std::vector<size_t>  CornerNodesOfFace( size_t face_id ) const;  
-    virtual std::vector<size_t>  NodesConnectedTo( size_t node_id ) const;
-    virtual void        CornerNodes( std::vector<size_t>& ids ) const;
-    virtual void        MidSideNodes( std::vector<size_t>& ids ) const;
-    virtual void        CounterClockwiseNodes( std::vector<size_t>& ids ) const;
-    virtual size_t      MidSideNodes() const { return 0; }
-    virtual size_t      CornerNodes() const  { return 4; }
+    virtual void        NodesOfSegment( uint32_t segm_id, std::vector<uint32_t>& snids ) const;
+    virtual void        NodesOfFace( uint32_t face_id, std::vector<uint32_t>& fnids ) const;
+    virtual std::vector<uint32_t>  CornerNodesOfFace( uint32_t face_id ) const;
+    virtual std::vector<uint32_t>  NodesConnectedTo( uint32_t node_id ) const;
+    virtual void        CornerNodes( std::vector<uint32_t>& ids ) const;
+    virtual void        MidSideNodes( std::vector<uint32_t>& ids ) const;
+    virtual void        CounterClockwiseNodes( std::vector<uint32_t>& ids ) const;
+    virtual uint32_t      MidSideNodes() const { return 0; }
+    virtual uint32_t      CornerNodes() const  { return 4; }
     virtual void        UnitNormal( std::vector<double>& vc ) const;
-    virtual void        UnitNormalToFace( size_t face, std::vector<double>& unrml ) const;
-    virtual CSMP_FEM_TYPE  ElementTypeOfFace( size_t face ) const;
-    virtual CSMP_FEM_TYPE  ElementTypeOfSegment( size_t /* segment */ ) const { return ISOPARAMETRIC_LINEAR_BAR; };
-    virtual void        ConsecutiveNodesAtBoundary( const std::vector<size_t>& bnodes,
-                                                   std::vector<size_t>& fnids );
+    virtual void        UnitNormalToFace( uint32_t face, std::vector<double>& unrml ) const;
+    virtual CSMP_FEM_TYPE  ElementTypeOfFace( uint32_t face ) const;
+    virtual CSMP_FEM_TYPE  ElementTypeOfSegment( uint32_t /* segment */ ) const { return ISOPARAMETRIC_LINEAR_BAR; };
+    virtual void        ConsecutiveNodesAtBoundary( const std::vector<uint32_t>& bnodes,
+                                                   std::vector<uint32_t>& fnids );
 
-    virtual   void      IntegrationPoint( size_t i, std::vector<double>& xyz ) const;
-    virtual double      WeightAtIntegrationPoint( size_t i ) const;
+    virtual   void      IntegrationPoint( uint32_t i, std::vector<double>& xyz ) const;
+    virtual double      WeightAtIntegrationPoint( uint32_t i ) const;
 
     virtual void        N( std::vector<double>& N, const std::vector<double>& xyz );
-    virtual void        N_AtIntegrationPoint( size_t ip, std::vector<double>& N );
+    virtual void        N_AtIntegrationPoint( uint32_t ip, std::vector<double>& N );
     virtual void        N_AtBaryCenter( std::vector<double>& N );
 
     virtual double      dN( DenseMatrix<DM_MIN>& dn, const std::vector<double>& xyz  );
     virtual void        dN( DenseMatrix<DM_MIN>& M );
-    virtual double      dN_AtIntegrationPoint( DenseMatrix<DM_MIN>& M, size_t gauss_point );
-    virtual double      dN_AtNode( DenseMatrix<DM_MIN>& M, size_t node );
+    virtual double      dN_AtIntegrationPoint( DenseMatrix<DM_MIN>& M, uint32_t gauss_point );
+    virtual double      dN_AtNode( DenseMatrix<DM_MIN>& M, uint32_t node );
     virtual double      dN_AtBarycenter( DenseMatrix<DM_MIN>& M );
 
-    virtual void        JacobianAtIntegrationPoint( size_t ip );
+    virtual void        JacobianAtIntegrationPoint( uint32_t ip );
     virtual double      JacobianInverse();
     virtual void        JacobianAt( const std::vector<double>& rst );
     virtual double      JacobianDeterminant();
@@ -55,7 +55,7 @@ public:
     virtual void        dNr( double r, double s, std::vector<double>& dNr ) const;
     virtual void        dNs( double r, double s, std::vector<double>& dNs ) const;
 
-    virtual   void      ExtrapolateIntegrationPointVariableToNodes( size_t nvars,
+    virtual   void      ExtrapolateIntegrationPointVariableToNodes( uint32_t nvars,
                                                                     const std::vector<double>& IVAR,
                                                                     std::vector<double>& NVAR ) const;
 
@@ -78,9 +78,9 @@ public:
     std::vector<double>         LXY;
     std::vector<double>         RS;
     bool                          use2Dto3Djacobi;
-    static const size_t           parametricDimehsions=2;
+    static const uint32_t           parametricDimehsions=2;
 
-    void Dimensions( size_t dimensions );
+    void Dimensions( uint32_t dimensions );
 
     double  Jacobi( const std::vector<double>& rs, std::vector<double>& EFG,
                        DenseMatrix<DM_MIN>& J );
