+++ source/includes/integration/finite_volumes/DES_transport_scheme/DES2PhaseTransport.h
@@ -9,12 +9,12 @@
 
 namespace csmp {
 
-template<size_t> class PropertyDatabase;
-template<size_t> class Region;
-template<size_t> class Model;
+template<uint32_t> class PropertyDatabase;
+template<uint32_t> class Region;
+template<uint32_t> class Model;
 
 
-template<size_t dim, template<size_t> class FLOW_FUNCTIONS>
+template<uint32_t dim, template<uint32_t> class FLOW_FUNCTIONS>
 class DES2PhaseTransport {
 
   public:
